backend/requirements.txt
# ScamShield Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai-whisper>=20231117
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.4.0
torch>=2.0.0
soundfile>=0.12.0
